refactor: combine `counter` and `time-series` into `/pages/[metricType]/[metricId].tsx` to eliminate duplicated code
